Key ASG name to its launch config
When an autoscaling-group's launch configuration changes those changes
aren't applied to any existing instances. This means that instances will
have to be manually terminated in order for the changes to take affect.

By keying the ASG to the launch configuration's name, though, we can
force AWS to terminate instances for us and bring up new ones with the
newly specified launch configuration.