Rough sketch to removing old nodes as adhoc
This allows the Puppetmaster to query what instances are running in the
estate (and other things that we should discuss).

It is possible to query the nodes in puppetdb using `govuk_node_list`,
although this does require a manual DNS entry in `/etc/hosts`.

The idea is to compare what is in puppetdb and what the API states in
running in the stack, and to remove a node if it doesn't report back a
match from the API.

I do not think this should be the long term way to remove nodes, but it
is a quick and dirty way to complete the task while we're testing.