Run build tool from anywhere
This commit adds a new argument (`GOVUK_AWS_REPO`) which you can use to
specify the location of this repo. It exchanges relative paths for
absolute paths, which gives the user the option of running the script
from anywhere.

I've submitted this commit because I'd prefer to have the script in
`/usr/local/bin` and run from whatever directory I'm currently working
from.

I've also changed the name because I love a tool with a snappy name.