Create frontend swarm cluster
This creates the "docker-frontend" project, which includes:

An ASG comprised of 2 workers
Three ASGs of a single manager each

Each manager has a static IP attached. See here for details about a
Swarm: https://docs.docker.com/engine/swarm/admin_guide/#add-manager-nodes-for-fault-tolerance

Managers are special and we should be able to know their IP - this is a
limitation of Docker Swarm itself.

They are joined to the same security group that allows full cluster
communications.

The idea is that a working Docker Swarm cluster could entirely replace
our frontend application instances (currently frontend,
calculators-frontend and content-store).

If a machine goes away, then it should hopefully be able to rejoin the
cluster and continue running the services it requires in that
environment.