Seperate internal services from public services
The `infra-public-services` project contained two different purposes:

1. Create top level internal DNS records that the application level
infrastructure is entirely dependent on. For example, if an application
wishes to speak to the current live Redis instance, it will query the
host "redis" which will resolve the currently live (green or blue) load
balancer for this service.

2. Public facing load balancers that resolve the publishing domain,
contain the public SSL certifcate, which are attached to the relevant
live autoscaling group, and then have a external DNS record assigned to
it. This is something that needs to be run when everything else is live
so that we can serve the infrastructure off the publishing API.

This splits out this project into two defined projects. It ensures that
the infra-internal-services can be updated and run without having to
touch or create any public load balancers. This is essentially just
creating or amending DNS records so should be a non-intrusive change.