Add environment common data
Update the build tool to allow the creation of environment common data:
data which is common across all environments, which should be stored in
`data/app-awesome-thing/common.tfvars`.

As a specific implementation this should *not* include secrets because
these should be different for each environment, and this should be
enforced.